+++ server/lib/python/cartodb_services/cartodb_services/tools/normalize.py
@@ -1,3 +1,6 @@
 def normalize(str_input):
     return str_input.replace('&quot;', '"') \
                     .replace(';', ',')
+
+def options_to_dict(options):
+    return dict(option.split("=") for option in options)